refactor(pool-page): extract _lib/{constants,helpers,tooltips,types}
Move pure declarations out of page.tsx into _lib/ subdir (Next.js App
Router excludes underscore-prefixed dirs from routing):

- _lib/constants.ts — TABS, SEARCH_PARAM_BY_TAB, MAX_TAB_LIMIT,
  MIN_REWARD_SAMPLE_SIZE
- _lib/helpers.ts — addressSearchTerms, matchesRowSearch, getTabLabel,
  getDebtTokenSideLabel, selectActiveOlsPool, decodePoolId, parseTabLimit
- _lib/tooltips.ts — BOUNDARY_TOOLTIP, EFFECTIVENESS_TOOLTIP,
  REWARD_TOOLTIP
- _lib/types.ts — OracleSortCol

page.tsx re-exports decodePoolId, parseTabLimit, getDebtTokenSideLabel,
selectActiveOlsPool so __tests__/ols.test.ts and page.test.tsx imports
from "../page" stay valid (pinned by exports.test.ts from prior commit).

Behavior-preserving — verified via:
- pnpm test (1366/1366 green, same as pre-refactor)
- pnpm typecheck (clean)
- chrome-devtools snapshot diff on /pool/<id>?tab=providers vs baseline:
  zero structural changes, only live-data drift (prices, timestamps)

page.tsx: 2831 → 2746 lines (-85, -3%).
