+++ ui-dashboard/src/app/pool/[poolId]/_lib/types.ts
@@ -1,0 +1,5 @@
+export type OracleSortCol =
+  | "timestamp"
+  | "oracleOk"
+  | "oraclePrice"
+  | "priceDifference";
